Port certs and restorable key to backup module
Both the RestorableRpmKey and the PEMCert classes got migrated to their
own module inside the backup folder.
